+++ src/elements/radio-button/radio-button-group/readme.md
@@ -77,7 +77,7 @@ In order to ensure readability for screen-readers, please provide an `aria-label
 | `orientation`         | `orientation`           | public  | `SbbOrientation`                                          | `'horizontal'` | Radio group's orientation, either horizontal or vertical. |
 | `radioButtons`        | -                       | public  | `(SbbRadioButtonElement \| SbbRadioButtonPanelElement)[]` |                | List of contained radio buttons.                          |
 | `required`            | `required`              | public  | `boolean`                                                 | `false`        | Whether the radio group is required.                      |
-| `size`                | `size`                  | public  | `SbbRadioButtonSize`                                      | `'m'`          | Size variant, either m or s.                              |
+| `size`                | `size`                  | public  | `SbbRadioButtonSize`                                      | `'m'`          | Size variant.                                             |
 | `value`               | `value`                 | public  | `any \| null \| undefined`                                |                | The value of the radio group.                             |
 
 ## Events
