+++ src/elements/radio-button/radio-button/radio-button.ts
@@ -1,9 +1,13 @@
 import type { CSSResultGroup, TemplateResult } from 'lit';
 import { LitElement, html, nothing } from 'lit';
-import { customElement } from 'lit/decorators.js';
+import { customElement, property } from 'lit/decorators.js';
 
 import { slotState } from '../../core/decorators.js';
-import { SbbRadioButtonCommonElementMixin, radioButtonCommonStyle } from '../common.js';
+import {
+  SbbRadioButtonCommonElementMixin,
+  radioButtonCommonStyle,
+  type SbbRadioButtonSize,
+} from '../common.js';
 
 import radioButtonStyle from './radio-button.scss?lit&inline';
 
@@ -19,6 +23,18 @@ export class SbbRadioButtonElement extends SbbRadioButtonCommonElementMixin(LitE
   public static readonly events = {
     stateChange: 'stateChange',
   } as const;
+
+  /**
+   * Size variant.
+   */
+  @property({ reflect: true })
+  public set size(value: SbbRadioButtonSize) {
+    this._size = value;
+  }
+  public get size(): SbbRadioButtonSize {
+    return this.group?.size ?? this._size;
+  }
+  private _size: SbbRadioButtonSize = 'm';
 
   protected override render(): TemplateResult {
     return html`
